WMSDK-518: Correcting code review comments
